pyproject.toml
[project]
name = "chinook-data-speech-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.0",
    "pyprojroot>=0.3.0",
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
]
